backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-google-genai
scikit-learn
numpy
python-multipart
